feature/RHP_174-make-leader-standings-only-query-current-season-records
Abstracted season controller methods to services using promises
